fix: поменять тип id пользователя в методе getUserEvents на int, мелкие правки по форматированию

+++ src/main/java/ru/yandex/practicum/filmrate/dao/EventStorage.java
@@ -5,7 +5,7 @@ import ru.yandex.practicum.filmrate.model.Event;
 import java.util.List;
 
 public interface EventStorage {
-	List<Event> getUserEvents(long id);
+	List<Event> getUserEvents(int id);
 
 	void addEvent(Event event);
 }
